% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{create.conda}
\alias{create.conda}
\title{Create conda environment with specified name.}
\usage{
create.conda(conda.env = "interactiontransformer")
}
\arguments{
\item{conda.env}{Name of conda environment.}
}
\description{
Create conda environment with specified name.
}
