% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{detect.conda}
\alias{detect.conda}
\title{Search for existing conda environment.}
\usage{
detect.conda(conda.env = "interactiontransformer")
}
\arguments{
\item{conda.env}{Name of conda environment.}
}
\description{
Search for existing conda environment.
}
