% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{get.top.interactions}
\alias{get.top.interactions}
\title{Given set of SHAP interaction values extracted from model, extract top interactions.}
\usage{
get.top.interactions(transformer)
}
\arguments{
\item{shap_vals}{List of SHAP values for each testing instance.}
}
\value{
DataFrame containing top interactions.
}
\description{
Given set of SHAP interaction values extracted from model, extract top interactions.
}
