% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{import_transformer}
\alias{import_transformer}
\title{Import interaction transformer package after sourcing python.}
\usage{
import_transformer()
}
\description{
Import interaction transformer package after sourcing python.
}
