% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{install_transformer}
\alias{install_transformer}
\title{Install InteractionTransformer Python Package.}
\usage{
install_transformer(
  custom.path = "git+https://github.com/jlevy44/InteractionTransformer",
  pip = F,
  conda.env = "interactiontransformer"
)
}
\arguments{
\item{custom.path}{Location/name of pip package.}

\item{pip}{Whether to use PyPI.}

\item{conda.env}{Name of conda environment to save to.}
}
\description{
This function installs the interaction transformer package from PyPI and places it in local or conda environment.
}
