% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{interaction.fit}
\alias{interaction.fit}
\title{Generate design matrix acquired from using SHAP on tree model.}
\usage{
interaction.fit(
  X.train,
  y.train,
  untrained_model = "default",
  max_train_test_samples = 100,
  mode_interaction_extract = "knee",
  include_self_interactions = F,
  cv_splits = 5L
)
}
\arguments{
\item{X.train}{Predictors in the form of a dataframe.}

\item{y.train}{One column dataframe containing outcomes.}

\item{untrained_model}{Scikit-learn tree-based estimator that has not been fit to the data.}

\item{max_train_test_samples}{Number of samples to train SHAP model off of.}

\item{mode_interaction_extract}{: Options for choosing number of interactions are 'sqrt' for square root number features, 'knee' for experimental knee method based on interaction scores, and any integer number of interactions.}

\item{include_self_interactions}{Whether to include self-interactions / quadratic terms.}

\item{cv_splits}{Number of CV splits for finding top interactions.}
}
\value{
Interaction transformer object that can be applied to the design matrix.
}
\description{
Returns transformer object. Y must be a single column dataframe.
}
