% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{interaction.transform}
\alias{interaction.transform}
\title{Using transformer, transform input design matrix.}
\usage{
interaction.transform(transformer, X)
}
\arguments{
\item{transformer}{Interaction Transformer object.}

\item{X}{Input design matrix.}
}
\value{
Output design matrix with interaction terms.
}
\description{
Using transformer, transform input design matrix.
}
