% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{refit.transformer}
\alias{refit.transformer}
\title{Modify the chosen design terms using a new number of top interactions extracted.}
\usage{
refit.transformer(transformer, new_number_interactions = 0L)
}
\arguments{
\item{new_number_interactions}{Number of new interactions to extract.}
}
\value{
Interaction transformer object that can be applied to the design matrix.
}
\description{
Modify the chosen design terms using a new number of top interactions extracted.
}
