% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{run.shap}
\alias{run.shap}
\title{Executes a SHAP routine over training and testing design matrices given the specified model and generates useful plots.}
\usage{
run.shap(
  X.train,
  X.test,
  model,
  model_type = "tree",
  savefile = "shap.output.png"
)
}
\arguments{
\item{X.train}{Training design matrix.}

\item{X.test}{Testing design matrix.}

\item{model}{Fit model.}

\item{model_type}{SHAP model type, choose between tree, kernel and linear.}

\item{savefile}{Save file for SHAP output image.}
}
\description{
Executes a SHAP routine over training and testing design matrices given the specified model and generates useful plots.
}
