% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{source.python}
\alias{source.python}
\title{Source python from user specified path.}
\usage{
source.python(python.exec = "/usr/local/bin/python3")
}
\arguments{
\item{python.exec}{Python executable.}
}
\description{
Source python from user specified path.
}
