% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_transformer.R
\name{train.test.split}
\alias{train.test.split}
\title{Split data into training and testing datasets. Y must be a single column dataframe.}
\usage{
train.test.split(X, y, train_size = 0.8)
}
\arguments{
\item{X}{Predictors in the form of a dataframe.}

\item{y}{One column dataframe containing outcomes.}

\item{train_size}{Fraction of dataset to split by, retains this fraction as training set.}
}
\value{
A list containing training and testing design matrices and outcome data.
}
\description{
Split data into training and testing datasets. Y must be a single column dataframe.
}
